feat: JSON 응답 값 구현
